+++ python/metatensor_core/metatensor/_c_api.py
@@ -157,6 +157,7 @@ DLPackDLTensorFromPyObjectNoSync = CFUNCTYPE(ctypes.c_int, ctypes.c_void_p, POIN
 DLPackCurrentWorkStream = CFUNCTYPE(ctypes.c_int, DLDeviceType, ctypes.c_int32, POINTER(POINTER(None)))
 DLPackManagedTensorToPyObjectNoSync = CFUNCTYPE(ctypes.c_int, POINTER(DLManagedTensorVersioned), POINTER(POINTER(None)))
 mts_data_origin_t = ctypes.c_uint64
+mts_create_mmap_array_callback_t = CFUNCTYPE(mts_status_t, ctypes.c_void_p, POINTER(c_uintptr_t), c_uintptr_t, DLDataType, c_uintptr_t, POINTER(mts_array_t))
 mts_realloc_buffer_t = CFUNCTYPE(ctypes.c_char_p, ctypes.c_void_p, ctypes.c_char_p, c_uintptr_t)
 mts_create_array_callback_t = CFUNCTYPE(mts_status_t, POINTER(c_uintptr_t), c_uintptr_t, DLDataType, POINTER(mts_array_t))
 
@@ -511,6 +512,13 @@ def setup_functions(lib):
     ]
     lib.mts_labels_load.restype = POINTER(mts_labels_t)
 
+    lib.mts_labels_load_mmap.argtypes = [
+        ctypes.c_char_p,
+        mts_create_mmap_array_callback_t,
+        ctypes.c_void_p,
+    ]
+    lib.mts_labels_load_mmap.restype = POINTER(mts_labels_t)
+
     lib.mts_labels_load_buffer.argtypes = [
         ctypes.c_char_p,
         c_uintptr_t,
@@ -545,6 +553,13 @@ def setup_functions(lib):
     ]
     lib.mts_block_load_buffer.restype = POINTER(mts_block_t)
 
+    lib.mts_block_load_mmap.argtypes = [
+        ctypes.c_char_p,
+        mts_create_mmap_array_callback_t,
+        ctypes.c_void_p,
+    ]
+    lib.mts_block_load_mmap.restype = POINTER(mts_block_t)
+
     lib.mts_block_save.argtypes = [
         ctypes.c_char_p,
         POINTER(mts_block_t),
@@ -572,6 +587,13 @@ def setup_functions(lib):
         mts_create_array_callback_t,
     ]
     lib.mts_tensormap_load_buffer.restype = POINTER(mts_tensormap_t)
+
+    lib.mts_tensormap_load_mmap.argtypes = [
+        ctypes.c_char_p,
+        mts_create_mmap_array_callback_t,
+        ctypes.c_void_p,
+    ]
+    lib.mts_tensormap_load_mmap.restype = POINTER(mts_tensormap_t)
 
     lib.mts_tensormap_save.argtypes = [
         ctypes.c_char_p,