+++ python/metatensor_core/metatensor/_labels.py
@@ -1086,7 +1086,6 @@ def _normalize_names_type(names: Union[str, Sequence[str]]) -> List[str]:
 def _create_new_labels(
     lib, names: List[str], values: np.ndarray, *, assume_unique: bool = False
 ):
-
     c_names = ctypes.ARRAY(ctypes.c_char_p, len(names))()
     for i, n in enumerate(names):
         c_names[i] = n.encode("utf8")